Fixed Plugin Loading. It all works now.
There was an issue with the way I loaded classes which lead to classes inheriting another class which hadn't been loaded throwing out errors. There's now a queue system in the idea that if a class is not present for a load but is in the queue, it will load. Generally the plugin loader should be more prepared for more complex plugins and should be a tad more organised for the eventual move over to a multi class loader plugin loader.